Move tests to their own folder. Use tox and add Travis CI config
