+++ tox.ini
@@ -1,0 +1,6 @@
+# content of: tox.ini , put in same dir as setup.py
+[tox]
+envlist = py27,py36
+[testenv]
+deps=pytest
+commands=py.test
